Implementing Layout for Desktop and fixing some minor issues
