+++ fe-library-app/src/components/AppRouter/pages/HomePage.tsx
@@ -1,22 +1,13 @@
-import { useState } from 'react'
-
-import Search from '../../Search/Search'
-import Filter from '../../Filter/Filter'
 import styles from './HomePage.module.css'
+import BookSearch from '../../BookSearch/BookSearch'
 
 const HomePage = () => {
-  const [ showFilter, setShowFilter ] = useState<boolean>(false)
-
-  const onFilterHandler = () => {
-    setShowFilter((pervState) => !pervState)
-  }
   return (
     <div className={styles.home}>
-      <Search />
-      <button onClick={onFilterHandler}>Filter</button>
-      <button>^</button>
-      {showFilter && <Filter/>}
-      <div>Home</div>
+      <div className={styles.screen_search}>
+        <BookSearch />
+      </div>
+      <div className={styles.home_content}>Home</div>
     </div>
   )
 }
