+++ fe-library-app/src/components/BookSearch/BookSearch.module.css
@@ -1,0 +1,29 @@
+.search button {
+    margin-left: 4px;
+    background-color: var(--dark_color);
+    color:white;
+    border-radius: 30px;
+    height: 30px;
+}
+
+.search a {
+    right: 10px;
+    position: fixed;
+    color: var(--dark_color)
+}
+
+@media only screen and (min-width: 768px){
+    .search{
+        position: absolute;
+        left: 15%;
+        right: 0;
+    }
+   .search button{
+        width: 40px;
+        height: 35px;
+   }
+   .search input{
+        height: 35px;
+        width: 40%;
+   }
+}
