+++ fe-library-app/src/components/Search/Search.module.css
@@ -1,4 +1,5 @@
 .search input {
     height: 30px;
     border-radius: 30px;
+    padding-left: 5px;
 }
