+++ fe-library-app/src/components/SideBar/SideBar.module.css
@@ -6,11 +6,25 @@
     min-height: 100vh;
     width: 50%;
     z-index: 90;
-    background-color: #282c34;
+    background-color: var(--dark_color);
     color:white;
     opacity: 0.9;
 }
 
 .sideBarHidden {
     display: none;
+}
+
+@media only screen and (min-width: 768px){
+    .sideBar{
+    display: block;
+    top: 25%;
+    bottom: 0;
+    left: 0;
+    position: fixed;
+    min-height: 100vh;
+    width: var(--sidebar_width);
+    background-color: var(--dark_color);
+    color:white;
+    }
 }
