+++ fe-library-app/src/variables.css
@@ -1,0 +1,5 @@
+body{
+    --sidebar_width: 15vw;
+    --dark_color: #282c34;
+    --orange_color: #F28741;
+}
